+++ Sheepskin/Library/Data/Instrument/Instrument/Instrument.cpp
@@ -12,6 +12,6 @@ void Instrument::addRecord(const Record &record) {
     records.push_back(record);
 }
 
-double Instrument::getVolatility(const Record &record) {
+long double Instrument::getVolatility(const Record &record) {
     return 0.0;
 }
